requirements.txt
streamlit==1.43.1
Pillow